refactored tooltips to be a widget file
