+++ express/code/blocks/simplified-pricing-cards/simplified-pricing-cards.css
@@ -311,85 +311,6 @@ main .section>.simplified-pricing-cards-wrapper {
     border-radius: 24px;
 }
 
-/* Tooltip CSS */
-.simplified-pricing-cards .tooltip  button.hover .tooltip-text,
-.simplified-pricing-cards .tooltip > button:focus-within .tooltip-text,
-.simplified-pricing-cards .tooltip-text.hover {
-    opacity: 1;
-    pointer-events: all;
-}
-
-.simplified-pricing-cards .tooltip-text {
-    opacity: 0;
-    padding: 8px;
-    z-index: 10;
-    background-color: var(--color-gray-800);
-    color: var(--color-white);
-    border-radius: 8px;
-    position: absolute;
-    bottom: 100%;
-    width:  200px;
-    left: -105px; 
-    margin-bottom: 5px;
-    user-select: text;
-    pointer-events: all;
-    transition: bottom 0.5s;
-}
-
-.simplified-pricing-cards .tooltip>button {
-    border: none;
-    background-color: transparent;
-    padding: 0px;
-    font-family: "adobe-clean";
-    position: relative;
-  }
-
-.simplified-pricing-cards .tooltip-text.overflow-left {
-    right: inherit;
-    left: -20px;
-}
-
-.simplified-pricing-cards .tooltip-text.overflow-right {
-    left: inherit;
-    right: -20px;
-}
-
-.simplified-pricing-cards .tooltip-text::after {
-    content: '';
-    position: absolute;
-    bottom: -5px;
-    left: 107px; 
-    border-top-color: var(--color-black);
-    border-left: 5px solid transparent;
-    border-right: 5px solid transparent;
-    border-top: 5px solid black;
-}
-
-.simplified-pricing-cards .tooltip-text.overflow-left::after {
-    right: inherit;
-    left: 23px;
-}
-
-.simplified-pricing-cards .tooltip-text.overflow-right::after {
-    left: inherit;
-    right: 20px;
-}
-
-
-.simplified-pricing-cards .tooltip-icon {
-    position: relative;
-    top: 2px;
-}
-
-.simplified-pricing-cards .tooltip {
-    position: relative;
-    display: inline;
-}
-
-.simplified-pricing-cards .tooltip span {
-    position: relative;
-}
-
 .simplified-pricing-cards .card-wrapper .card.hide .plan-explanation,
 .simplified-pricing-cards .card-wrapper .card.hide .pricing-area,
 .simplified-pricing-cards .card-wrapper .card.hide .card-cta-group {
