+++ express/code/blocks/simplified-pricing-cards/simplified-pricing-cards.js
@@ -1,9 +1,10 @@
-import { getLibs, getIconElementDeprecated, addTempWrapperDeprecated, decorateButtonsDeprecated } from '../../scripts/utils.js';
+import { getLibs,addTempWrapperDeprecated, decorateButtonsDeprecated } from '../../scripts/utils.js';
 import {
   fetchPlanOnePlans,
   formatDynamicCartLink,
 } from '../../scripts/utils/pricing.js';
 import { debounce } from '../../scripts/utils/hofs.js';
+import handleTooltip, { adjustElementPosition } from '../../scripts/widgets/tooltip.js';
 
 let createTag; let getConfig;
 let replaceKeyArray; let formatSalesPhoneNumber;
@@ -11,86 +12,6 @@ let replaceKeyArray; let formatSalesPhoneNumber;
 const SALES_NUMBERS = '((business-sales-numbers))';
 const PRICE_TOKEN = '((pricing))';
 const YEAR_2_PRICING_TOKEN = '((year-2-pricing-token))';
-
-export function adjustElementPosition() {
-  const elements = document.querySelectorAll('.tooltip-text');
-
-  if (elements.length === 0) return;
-  for (const element of elements) {
-    const rect = element.getBoundingClientRect();
-    if (rect.right > window.innerWidth) {
-      element.classList.remove('overflow-left');
-      element.classList.add('overflow-right');
-    } else if (rect.left < 0) {
-      element.classList.remove('overflow-right');
-      element.classList.add('overflow-left');
-    }
-  }
-}
-
-export function handleTooltip(pricingArea) {
-  const elements = pricingArea.querySelectorAll('p');
-  const tooltipPattern = /\(\(([^]+)\)\)([^]+)\(\(\/([^]+)\)\)/g;
-  let tooltipMatch;
-  let tooltipContainer;
-
-  Array.from(elements).forEach((p) => {
-    const match = tooltipPattern.exec(p.textContent);
-    if (match) {
-      tooltipMatch = match;
-      tooltipContainer = p;
-    }
-  });
-  if (!tooltipMatch) return;
-
-  tooltipContainer.innerHTML = tooltipContainer.innerHTML.replace(tooltipPattern, '');
-  const tooltipContent = tooltipMatch[2];
-  tooltipContainer.classList.add('tooltip');
-  
-  const tooltipPopup = createTag('div', { class: 'tooltip-text' });
-  tooltipPopup.innerText = tooltipContent;
-  
-  const infoIcon = getIconElementDeprecated('info', 44, 'Info', 'tooltip-icon');
-  const tooltipButton = createTag('button');
-  tooltipButton.setAttribute('aria-label', tooltipContent);
-  infoIcon.setAttribute('tabindex', 0);
-  
-  tooltipButton.append(infoIcon);
-  tooltipButton.append(tooltipPopup);
-  tooltipContainer.append(tooltipButton);
-  
-  tooltipButton.addEventListener('click', adjustElementPosition);
-  window.addEventListener('resize', adjustElementPosition);
-
-  infoIcon.addEventListener('mouseover', () => {
-    tooltipButton.classList.add('hover');
-  });
-
-  infoIcon.addEventListener('mouseleave', () => {
-    setTimeout(() => {
-      tooltipButton.classList.remove('hover');
-    }, 500);
-  });
-
-  tooltipPopup.addEventListener('mouseover', () => {
-    if (tooltipButton.classList.contains('hover')) {
-      tooltipPopup.classList.add('hover');
-    }
-  });
-
-  tooltipPopup.addEventListener('mouseleave', () => {
-    setTimeout(() => {
-      tooltipPopup.classList.remove('hover');
-    }, 500);
-  });
-
-  document.addEventListener('keydown', (e) => {
-    if (e.key === 'Escape') {
-      document.activeElement.blur();
-      tooltipPopup.classList.remove('hover');
-    }
-  });
-}
 
 function getHeightWithoutPadding(element) {
   const styles = window.getComputedStyle(element);
